+++ test/pruning_1.x_v1/test_pattern_lock.py
