+++ test/pruning_1.x_v2/test_pruning_patterns.py
@@ -6,47 +6,46 @@ import torch.nn as nn
 
 from neural_compressor.data import Datasets
 from neural_compressor.experimental.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
+from neural_compressor.conf.pythonic_config import Config
 from neural_compressor.config import WeightPruningConfig
-from neural_compressor.pruner.pruning import Pruning
+from neural_compressor.experimental.pruning_v2 import Pruning
 
 
-class TestPruningCriteria(unittest.TestCase):
+class TestPruningPatterns(unittest.TestCase):
     model = torchvision.models.resnet18()
 
-    def test_pruning_criteria(self):
+    def test_pruning_pattern(self):
         local_configs = [
             {
                 "op_names": ['layer1.*'],
-                'target_sparsity': 0.4,
-                "pattern": '8x2',
-                "pruning_type": "magnitude_progressive",
-                "pruning_scope": "local",
-                "sparsity_decay_type": "cube"
+                'target_sparsity': 0.5,
+                "pattern": '5:8',
+                "pruning_type": "magnitude"
             },
             {
                 "op_names": ['layer2.*'],
-                'target_sparsity': 0.45,
-                'pattern': '2:4',
-                "pruning_type": "snip",
-                'start_step': 6,
-                'end_step': 6
+                "pattern": '1xchannel',
+                "pruning_scope": "global"
             },
             {
+                "start_step": 2,
+                "end_step": 20,
                 "op_names": ['layer3.*'],
-                'excluded_op_names': ['downsample.*'],
-                'target_sparsity': 0.7,
-                'pattern': '4x1',
-                "pruning_type": "snip_momentum_progressive",
-                "pruning_frequency": 4,
-                "min_sparsity_ratio_per_op": 0.5,
-                "max_sparsity_ratio_per_op": 0.8,
+                'target_sparsity': 0.666666,
+                'pattern': '4x2',
+                "pruning_type": "snip_progressive",
+                "pruning_frequency": 5
             }
         ]
-        config = WeightPruningConfig(
+        conf = WeightPruningConfig(
             local_configs,
             target_sparsity=0.8,
-            sparsity_decay_type="cube"
+            sparsity_decay_type="cos",
+            excluded_op_names=["downsample.*"],
+            pruning_scope="local",
+            min_sparsity_ratio_per_op=0.1
         )
+        config = Config(quantization=None, benchmark=None, pruning=conf, distillation=None)
         prune = Pruning(config)
         prune.update_config(start_step=1, end_step=10)
         prune.model = self.model
@@ -58,8 +57,7 @@ class TestPruningCriteria(unittest.TestCase):
         dummy_dataloader = PyTorchDataLoader(dummy_dataset)
 
         prune.on_train_begin()
-        prune.update_config(pruning_frequency=4)
-        for epoch in range(2):
+        for epoch in range(5):
             self.model.train()
             prune.on_epoch_begin(epoch)
             local_step = 0
