+++ test/pruning_1.x_v2/test_pruning_schedulers.py
@@ -6,51 +6,46 @@ import torch.nn as nn
 
 from neural_compressor.data import Datasets
 from neural_compressor.experimental.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
+from neural_compressor.conf.pythonic_config import Config
 from neural_compressor.config import WeightPruningConfig
-from neural_compressor.pruner.pruning import Pruning
+from neural_compressor.experimental.pruning_v2 import Pruning
 
-local_types_config = [
+local_schedulers_config = [
     {
         "start_step": 0,
-        "end_step": 0,
-        "pruning_type": "pattern_lock",
+        "end_step": 2,
+        "pruning_type": "magnitude",
         "op_names": ['layer1.*'],
         "excluded_op_names": ['layer2.*'],
-        "pruning_scope": "global"
+        "pruning_scope": "global",
+        "target_sparsity": 0.5,
+        "pattern": "4x1"
     },
     {
         "start_step": 1,
-        "end_step": 1,
+        "end_step": 10,
         "target_sparsity": 0.5,
-        "pruning_type": "snip_momentum_progressive",
+        "pruning_type": "snip_momentum",
         "pruning_frequency": 2,
         "op_names": ['layer2.*'],
         "pruning_scope": "local",
-        "pattern": "4x1",
+        "target_sparsity": 0.75,
+        "pattern": "32x1",
         "sparsity_decay_type": "exp"
-    },
-    {
-        "start_step": 2,
-        "end_step": 8,
-        "target_sparsity": 0.8,
-        "pruning_type": "snip_progressive",
-        "pruning_frequency": 1,
-        "op_names": ['layer3.*'],
-        "pruning_scope": "local",
-        "pattern": "16x1",
-        "sparsity_decay_type": "cube"
     }
 ]
 
-fake_snip_config = WeightPruningConfig(local_types_config, target_sparsity=0.9, start_step=0, \
-                                       end_step=10, pruning_frequency=3, sparsity_decay_type="exp")
+fake_snip_config = WeightPruningConfig(local_schedulers_config, target_sparsity=0.9, start_step=0, \
+                                       end_step=10, pruning_frequency=1, sparsity_decay_type="exp")
 
 
-class TestPruningTypes(unittest.TestCase):
+class TestPruningCriteria(unittest.TestCase):
     model = torchvision.models.resnet18()
 
-    def test_pruning_types(self):
-        prune = Pruning(fake_snip_config)
+    def test_pruning_schedulers(self):
+        config = Config(quantization=None, benchmark=None, pruning=fake_snip_config, distillation=None)
+        prune = Pruning(config)
+        prune.update_config(start_step=1)
         prune.model = self.model
         criterion = nn.CrossEntropyLoss()
         optimizer = torch.optim.SGD(self.model.parameters(), lr=0.0001)
