+++ test/pruning_2.x/test_pruning_patterns.py
