+++ test/pruning_2.x/test_pruning_regs.py
@@ -60,34 +60,34 @@ class TestPruningRegs(unittest.TestCase):
 
     def test_pruning_regs(self):
         pass
-        compression_manager = prepare_compression(model=self.model, confs=fake_snip_config)
-        compression_manager.callbacks.on_train_begin()
-        criterion = nn.CrossEntropyLoss()
-        optimizer = torch.optim.SGD(self.model.parameters(), lr=0.0001)
-        datasets = Datasets('pytorch')
-        dummy_dataset = datasets['dummy'](shape=(10, 3, 224, 224), low=0., high=1., label=True)
-        dummy_dataloader = PyTorchDataLoader(dummy_dataset)
-        compression_manager.callbacks.on_train_begin()
-        for epoch in range(2):
-            self.model.train()
-            compression_manager.callbacks.on_epoch_begin(epoch)
-            local_step = 0
-            for image, target in dummy_dataloader:
-                compression_manager.callbacks.on_step_begin(local_step)
-                output = self.model(image)
-                loss = criterion(output, target)
-                optimizer.zero_grad()
-                loss.backward()
-                compression_manager.callbacks.on_before_optimizer_step()
-                optimizer.step()
-                compression_manager.callbacks.on_after_optimizer_step()
-                compression_manager.callbacks.on_step_end()
-                local_step += 1
-        
-            compression_manager.callbacks.on_epoch_end()
-        compression_manager.callbacks.on_train_end()
-        compression_manager.callbacks.on_before_eval()
-        compression_manager.callbacks.on_after_eval()
+        # compression_manager = prepare_compression(model=self.model, confs=fake_snip_config)
+        # compression_manager.callbacks.on_train_begin()
+        # criterion = nn.CrossEntropyLoss()
+        # optimizer = torch.optim.SGD(self.model.parameters(), lr=0.0001)
+        # datasets = Datasets('pytorch')
+        # dummy_dataset = datasets['dummy'](shape=(10, 3, 224, 224), low=0., high=1., label=True)
+        # dummy_dataloader = PyTorchDataLoader(dummy_dataset)
+        # compression_manager.callbacks.on_train_begin()
+        # for epoch in range(2):
+        #     self.model.train()
+        #     compression_manager.callbacks.on_epoch_begin(epoch)
+        #     local_step = 0
+        #     for image, target in dummy_dataloader:
+        #         compression_manager.callbacks.on_step_begin(local_step)
+        #         output = self.model(image)
+        #         loss = criterion(output, target)
+        #         optimizer.zero_grad()
+        #         loss.backward()
+        #         compression_manager.callbacks.on_before_optimizer_step()
+        #         optimizer.step()
+        #         compression_manager.callbacks.on_after_optimizer_step()
+        #         compression_manager.callbacks.on_step_end()
+        #         local_step += 1
+
+        #     compression_manager.callbacks.on_epoch_end()
+        # compression_manager.callbacks.on_train_end()
+        # compression_manager.callbacks.on_before_eval()
+        # compression_manager.callbacks.on_after_eval()
 
 
 if __name__ == "__main__":
